added a 'my profile' page, changed some code layout

+++ controllers/timetable.js
@@ -5,37 +5,36 @@ const members = require('../modules/member');
 const giveRoles = require('../bot/createClass');
 
 async function getTimetableForm(req, res) {
-	let pool = getDB();
-	// ensure that the user has signed up
-	if (!req.session.user) {
-		res.redirect('/login?redirect=/signup/timetable')
-		return;
-	}
-	user = await pool.query(`
+	const pool = getDB();
+	// ensure that the user is signed in
+	if (!req.session.user) 
+		return res.redirect('/login?redirect=/signup/timetable');
+	
+	const user = await pool.query(`
 		SELECT * FROM users
 		WHERE member_id = '${req.session.user.member_id}'
 		AND group_id('${req.session.user.guild_id}') = group_id
 		AND complete = FALSE;
-	`).catch(console.error)
-	if (user.rowCount == 0) {
-		res.redirect('/myprofile')
-		return;
-	}
+	`).catch(console.error);
+
+	if (user.rowCount == 0)
+		return res.redirect('/');
+
 	// get the sets and their divisions
-	let sets = await pool.query(`
-    SELECT	DISTINCT
-      sets.set 	        AS name,
-    	divisions.name 		AS division,
-    	teachers.name 		AS teacher,
-    	sets.name  		    AS subject
-    FROM sets
-    INNER JOIN divisions 	ON divisions.id = sets.division_id
-    INNER JOIN groups		  ON divisions.group_id = groups.id
-    INNER JOIN timetable	ON sets.id = timetable.set_id
-    LEFT JOIN  teachers		ON timetable.teacher_id = teachers.id
-    WHERE groups.guild_id = '${req.session.user.guild_id}' AND timetable.usual
-    ORDER BY division DESC;
-  `)
+	const sets = await pool.query(`
+		SELECT	DISTINCT
+			sets.set 	        AS name,
+			divisions.name 		AS division,
+			teachers.name 		AS teacher,
+			sets.name  		    AS subject
+		FROM sets
+		INNER JOIN divisions 	ON divisions.id = sets.division_id
+		INNER JOIN groups		ON divisions.group_id = groups.id
+		INNER JOIN timetable	ON sets.id = timetable.set_id
+		LEFT JOIN  teachers		ON timetable.teacher_id = teachers.id
+		WHERE groups.guild_id = '${req.session.user.guild_id}' AND timetable.usual
+		ORDER BY division DESC;
+	`).catch(console.error);
 	// render page with
 	res.render("pages/timetable", {
 		subjects: sets.rows,
@@ -46,38 +45,37 @@ async function getTimetableForm(req, res) {
 
 
 function getTimetable(req, res) {
-	//if the user isn't logged in, they cannot see the timetable data
-	if (!req.session.user) {
-		res.status(403).send('oof! log in to see this data!')
-		return;
-	}
-	let pool = getDB()
+	// if the user isn't logged in, they cannot see the timetable data
+	if (!req.session.user) 
+		return res.status(403).end();
+
+	const pool = getDB();
 	pool.query(`
 		SELECT
-      timetable.teacher,
-  		timetable.period,
-  		timetable.day,
-  		timetable.class,
-  		timetable.division,
-  		override.name      AS override
+			timetable.teacher,
+			timetable.period,
+			timetable.day,
+			timetable.class,
+			timetable.division,
+			override.name      AS override
 		FROM divisions AS override
 		RIGHT JOIN (
 			SELECT
-        teachers.name     AS teacher,
-		    timetable.period,
-        timetable.day,
-        subject.name      AS class,
-        divisions.name    AS division,
-        sets.overrides    AS overrides
-			FROM timetable
-			LEFT JOIN teachers  	ON teachers.id = timetable.teacher_id
-			INNER JOIN sets 	 		ON sets.id = timetable.set_id
-			INNER JOIN subject 	 	ON timetable.subject_id = subject.id
-			INNER JOIN divisions 	ON sets.division_id = divisions.id
-      INNER JOIN groups 		ON divisions.group_id = groups.id
+			teachers.name     AS teacher,
+			timetable.period,
+			timetable.day,
+			subject.name      AS class,
+			divisions.name    AS division,
+			sets.overrides    AS overrides
+		FROM timetable
+		LEFT JOIN teachers  	ON teachers.id = timetable.teacher_id
+		INNER JOIN sets 	 		ON sets.id = timetable.set_id
+		INNER JOIN subject 	 	ON timetable.subject_id = subject.id
+		INNER JOIN divisions 	ON sets.division_id = divisions.id
+		INNER JOIN groups 		ON divisions.group_id = groups.id
 			WHERE sets.set  = '${req.query.set}' AND
-      divisions.name  = '${req.query.sub}' AND
-      groups.guild_id = '${req.session.user.guild_id}') AS timetable
+			divisions.name  = '${req.query.sub}' AND
+			groups.guild_id = '${req.session.user.guild_id}') AS timetable
 		ON timetable.overrides = override.id;`)
 		.then((responce) => {
 			res.json(responce)
@@ -87,28 +85,34 @@ function getTimetable(req, res) {
 
 
 async function giveClasses(req, res) {
-	let pool = getDB();
+
+	if (!req.session.user) 
+		return res.status(400).end();
+
+	res.redirect('/my_profile');
+
+	const pool = getDB();
 	// store the user in their session
 	req.session.user = Object.assign(req.session.user, req.body);
-	let member = members.get(req.session.user.guild_id, req.session.user.member_id);
+	const member = members.get(req.session.user.guild_id, req.session.user.member_id);
 	// save new user's classes to database
 	await saveClasses(req.session.user);
 	// remove the pre-user
-	pool.query(`DELETE FROM pre_users
-    WHERE member_id = '${req.session.user.member_id}'
-    AND guild_id = '${req.session.user.guild_id}';
-  `).catch(console.error);
-	res.redirect(`/`);
+	await pool.query(`DELETE FROM pre_users
+		WHERE member_id = '${req.session.user.member_id}'
+		AND guild_id = '${req.session.user.guild_id}';
+	`).catch(console.error);
 	// give the user access to their classes.
 	await giveRoles(member, req.body.classes);
+	// DM the user on discord
 	member.send("All done! If you want me again then type `help`");
 }
 
 function saveClasses(user) {
-	let pool = getDB();
+	const pool = getDB();
 	let query = `
-  INSERT INTO usr_set_join
-  VALUES`
+		INSERT INTO usr_set_join
+		VALUES`
 	for (let division in user.classes) {
 		if (user.classes[division] != 'none') {
 			query += `
